SerialConnection and its interface renamed
Renamed to avoid compatibility issues
