+++ src/main/java/com/ghgande/j2mod/modbus/facade/ModbusSerialMaster.java
@@ -17,8 +17,8 @@ package com.ghgande.j2mod.modbus.facade;
 
 import com.ghgande.j2mod.modbus.Modbus;
 import com.ghgande.j2mod.modbus.io.AbstractModbusTransport;
-import com.ghgande.j2mod.modbus.net.JSerialCommPort;
 import com.ghgande.j2mod.modbus.net.SerialConnection;
+import com.ghgande.j2mod.modbus.net.SerialConnectionInterface;
 import com.ghgande.j2mod.modbus.util.SerialParameters;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -34,7 +34,7 @@ import org.slf4j.LoggerFactory;
 public class ModbusSerialMaster extends AbstractModbusMaster {
 
     private static final Logger logger = LoggerFactory.getLogger(ModbusSerialMaster.class);
-    private SerialConnection connection;
+    private SerialConnectionInterface connection;
 
     /**
      * Constructs a new master facade instance for communication
@@ -57,7 +57,7 @@ public class ModbusSerialMaster extends AbstractModbusMaster {
      */
     public ModbusSerialMaster(SerialParameters param, int timeout) {
         try {
-            connection = new JSerialCommPort(param);
+            connection = new SerialConnection(param);
             connection.setTimeout(timeout);
             this.timeout = timeout;
         }
