+++ src/main/java/com/ghgande/j2mod/modbus/io/ModbusSerialTransaction.java
@@ -21,8 +21,7 @@ import com.ghgande.j2mod.modbus.ModbusIOException;
 import com.ghgande.j2mod.modbus.ModbusSlaveException;
 import com.ghgande.j2mod.modbus.msg.ExceptionResponse;
 import com.ghgande.j2mod.modbus.msg.ModbusRequest;
-import com.ghgande.j2mod.modbus.net.JSerialCommPort;
-import com.ghgande.j2mod.modbus.net.SerialConnection;
+import com.ghgande.j2mod.modbus.net.SerialConnectionInterface;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -68,7 +67,7 @@ public class ModbusSerialTransaction extends ModbusTransaction {
      *
      * @param con a <tt>TCPMasterConnection</tt> instance.
      */
-    public ModbusSerialTransaction(SerialConnection con) {
+    public ModbusSerialTransaction(SerialConnectionInterface con) {
         setSerialConnection(con);
     }
 
@@ -76,9 +75,9 @@ public class ModbusSerialTransaction extends ModbusTransaction {
      * Sets the port on which this <tt>ModbusTransaction</tt>
      * should be executed.
      *
-     * @param con a <tt>JSerialCommPort</tt>.
+     * @param con a <tt>SerialConnection</tt>.
      */
-    public void setSerialConnection(SerialConnection con) {
+    public void setSerialConnection(SerialConnectionInterface con) {
         synchronized (MUTEX) {
             transport = con.getModbusTransport();
         }
