+++ src/test/java/com/ghgande/j2mod/modbus/net/ModbusMasterFactory.java
@@ -56,13 +56,13 @@ public class ModbusMasterFactory {
             parms.setPortName(parts[1]);
             parms.setBaudRate(9600);
             parms.setDatabits(8);
-            parms.setParity(SerialConnection.NO_PARITY);
+            parms.setParity(SerialConnectionInterface.NO_PARITY);
             parms.setStopbits(1);
-            parms.setFlowControlIn(SerialConnection.FLOW_CONTROL_DISABLED);
+            parms.setFlowControlIn(SerialConnectionInterface.FLOW_CONTROL_DISABLED);
             parms.setEcho(false);
             try {
                 ModbusRTUTransport transport = new ModbusRTUTransport();
-                transport.setCommPort(JSerialCommPort.getCommPort(parms.getPortName()));
+                transport.setCommPort(SerialConnection.getCommPort(parms.getPortName()));
                 transport.setEcho(false);
                 return transport;
             }
